python/api/pyproject.toml
[project]
name = "deadlock-api-client"
version = "0.1.0"
description = "A client library for accessing Deadlock API"
authors = []
requires-python = "~=3.9"
readme = "README.md"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0,<3",
]

[tool.uv.build-backend]
module-name = "deadlock_api_client"
module-root = ""
data = [
    "CHANGELOG.md",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
